Use multiple urls for health check
